\name{extarctBiweekMegteos}
\alias{extarctBiweekMegteos}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Extract the biweekly averages for the subject location and timeline }
\description{ Extract the biweekly averages for the subject location and timeline }
\usage{
extarctBiweekMegteos(dfInput = dfInput, nVar = "precip33", flag = "precip", from_biwk = 1, to_biwk = 10, subgid = "geoid_latlongt")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dfInput}{Input data frame}
  \item{nVar}{
  the output name for meteorological parameters:
      tmmn: daily minimum temperature;
      tmmx: daily maximum temperature;
      vs: wind speed;
      th: wind direction;
      sph: specific humidity ;
      srad: surface downwelling shortwave flux in air
  }
  \item{flag}{ the flag name for the file generated by  genBiweekMeteos}
  \item{from_biwk}{starting biweek id }
  \item{to_biwk}{ending biweek id}
  \item{subgid}{id }
}
\details{Extract the biweekly averages for the subject location and timeline }
\value{Data frame with the nVar variable added}
\author{Lianfa Li }
\examples{
#Extract the value of meteorological parameters based on the output of genBiweekMeteos,
# using the same flag genBiweekMeteos.  Here is the example for biweekly average wind speed.

  data(sc_sample)
  sc_sample$wndsp=extractBiweekMegteos(dfInput=sc_sample,nVar="vs",flag="sc_sample",from_biwk=1,
                               to_biwk=574,subgid="gid")

}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
